+++ react/javascript/src/index.ts
@@ -1,4 +1,3 @@
 import GherkinDocumentList from './components/app/GherkinDocumentList'
-import Wrapper from './components/app/Wrapper'
 import QueriesWrapper from './components/app/QueriesWrapper'
-export { GherkinDocumentList, Wrapper, QueriesWrapper }
+export { GherkinDocumentList, QueriesWrapper }